Marking InitializeStore() as obsolete. Closes GH-2185
