+++ src/Marten/DefaultSessionFactory.cs
@@ -1,0 +1,22 @@
+namespace Marten
+{
+    internal class DefaultSessionFactory: ISessionFactory
+    {
+        private readonly IDocumentStore _store;
+
+        public DefaultSessionFactory(IDocumentStore store)
+        {
+            _store = store;
+        }
+
+        public IQuerySession QuerySession()
+        {
+            return _store.QuerySession();
+        }
+
+        public IDocumentSession OpenSession()
+        {
+            return _store.OpenSession();
+        }
+    }
+}
