+++ src/Marten/ISessionFactory.cs
@@ -1,0 +1,21 @@
+namespace Marten
+{
+    /// <summary>
+    /// Pluggable strategy for customizing how IDocumentSession / IQuerySession
+    /// objects are created within an application.
+    /// </summary>
+    public interface ISessionFactory
+    {
+        /// <summary>
+        /// Build new instances of IQuerySession on demand
+        /// </summary>
+        /// <returns></returns>
+        IQuerySession QuerySession();
+
+        /// <summary>
+        /// Build new instances of IDocumentSession on demand
+        /// </summary>
+        /// <returns></returns>
+        IDocumentSession OpenSession();
+    }
+}
